feat(kernel): use mkmm classic file cp version instead of mounting tmpfs
looked at source code. now i think it is, actually, better to use
copying strategy to preserve kernel modules. because they are still on
the same partition and copied using hard links. which means, it actually
not copied at all. whereas it the case of using tmpfs partition, we
explicitly copy everything to ram only for a benefit to just unmount and
automatically flush the contents of in-memory partition.